since someone else leaked the keys, i'm making pulic this repo

+++ setup.py
@@ -20,15 +20,16 @@ if __name__ == "__main__":
         print("If you are unsure how to do this, please refer to the documentation: https://github.com/castlabs/electron-releases/wiki/EVS")
         print("If you do not sign the executable, video playback might not work on Windows and macOS.")
         print("="*90)
-        have_evs_acc = input("Do you have a Castlabs EVS account? (Note: If you say no, you will be asked to sign up.) ([y]/n): ")
+        have_evs_acc = input("Do you have a Castlabs EVS account? (Note: If you say no, you will be asked to sign up. Use 's' to skip.) ([y]/n/s): ")
 
         if have_evs_acc.lower() in ["n", "no"]:
             os.system(f"{sys.executable} -m castlabs_evs.account signup")
 
-        if sys.platform == "win32":
-            os.system(f"{sys.executable} -m castlabs_evs.vmp sign-pkg .\\electron-player\\node_modules\\electron\\dist")
-        elif sys.platform == "darwin":
-            os.system(f"{sys.executable} -m castlabs_evs.vmp sign-pkg ./electron-player/node_modules/electron/dist/")
+        if have_evs_acc.lower() not in ["s", "skip"]:
+            if sys.platform == "win32":
+                os.system(f"{sys.executable} -m castlabs_evs.vmp sign-pkg .\\electron-player\\node_modules\\electron\\dist")
+            elif sys.platform == "darwin":
+                os.system(f"{sys.executable} -m castlabs_evs.vmp sign-pkg ./electron-player/node_modules/electron/dist/")
 
     current_dir = os.getcwd()
 
@@ -60,6 +61,21 @@ if __name__ == "__main__":
         os.chdir(current_dir)
 
     if sys.platform == "linux":
-        os.system(f"pkexec apt-get install libportaudio2")
+        print("="*90)
+        print("WARNING: In order to cast screen, you must install the libportaudio2 package.")
+        print("="*90)
+        installNow = input("Do you want to install it now? ([y]/n): ")
+        if installNow.lower() in ["y", "yes", ""]:
+            # check package manager
+            if os.path.exists("/usr/bin/apt-get"):
+                os.system(f"pkexec apt-get install libportaudio2")
+            elif os.path.exists("/usr/bin/pacman"):
+                os.system(f"pkexec pacman -S libportaudio")
+            elif os.path.exists("/usr/bin/yum"):
+                os.system(f"pkexec yum install libportaudio2")
+            else:
+                print("="*90)
+                print("WARNING: Could not find a supported package manager. Please install libportaudio2 manually.")
+                print("="*90)
 
     os.system(f"{sys.executable} main.py")
